Redesign the config format
The config files used to be in the form `"key_x": [97, 98, 99]`, to signify
that the line in the DOSBox mapper for `key_x` should take exactly the
three scan codes 97, 98, and 99.  Lots of problems there:
 - It was confusing
 - I wrote config files with multiple bugs in them
 - Every key would send both its new value and its old value, meaning
   there needed to be lines like `"key_a": []` so the key `a` sent only
	 the value `x` and not also the original value `a`.  Keeping these lines
   in sync was a challenge.
 - Writing the files meant repeatedly looking up the mapping from arcade
   button (p1b1) to logical key name (lctrl) to scan code (306).

So I rewrote the config format to closely mirror the comments I'd been
writing and just derive all the rest.  Now each config line is like:
`"p1b1": "x"`.  The wrapper automatically looks up what scan code `p1b1`
is and tells DOSBox that it should send an `x` and stop sending its
default value (left control).